Remove JitState from public interface
It’s an internal type only.
